+++ src/components/Footer/Footer.scss
@@ -1,6 +1,6 @@
-@import "./../styles/Variables.scss";
-@import "./../styles/Fonts.scss";
-@import "./../styles/Variables.scss";
+@import "src/styles/Variables";
+@import "src/styles/Fonts";
+@import "src/styles/Variables";
 
 .footer {
   background-color: $black;
