+++ src/components/Invite/Components/Preview.js
@@ -4,7 +4,7 @@ import { Nav } from "react-bootstrap";
 import PreviewBordorSvg from "../../../assets/images/Invite/PreviewBordor.svg";
 import Header from "../../Header/Header";
 import PlayerWelcome from "../../PlayerWelcome/PlayerWelcome";
-import Footer from "../../Footer";
+import Footer from "../../Footer/Footer";
 import * as MdIcons from "react-icons/md";
 import "./Preview.scss";
