+++ src/components/Utility/Invite/Invite.scss
@@ -1,4 +1,4 @@
-@import "../styles/Fonts.scss";
+@import "src/styles/Fonts";
 
 .invite-wrapper {
   width: 100%;
