+++ src/components/Utility/Popup/Popup.scss
@@ -1,4 +1,4 @@
-@import "../styles/Components.scss";
+@import "src/styles/Components";
 
 .modal {
     margin-top: 10rem;
@@ -32,7 +32,7 @@
     display: flex;
     flex-direction: column;
     justify-content: flex-end;
-    
+
 }
 
 .modal p {
